requirements.txt
requests~=2.28.1
qrcode~=8.1
pillow
brotli